+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.8.0
+- Added support for custom masks
+- Fixed sometimes items use the wrong behaviour
+- Added more resource and asset bundle helpers
+
 ## 2.7.4
 - Fixed latch fix modifying the base info
 - Fixed stackable abilities activating twice when they shouldn't